code-agent/chunker/requirements.txt
# Code Chunker Dependencies
openai>=1.0.0
python-dotenv>=1.0.0
